Ready for test

+++ src/main/java/inu/thebite/toryaba/service/serviceImpl/ClassServiceImpl.java
@@ -22,9 +22,13 @@ public class ClassServiceImpl implements ClassService {
 
     @Transactional
     @Override
-    public Class addClass(Long centerId, ClassRequest classRequest) {
+    public Class addClass(String userId, Long centerId, ClassRequest classRequest) {
         Center center = centerRepository.findById(centerId)
                 .orElseThrow(() -> new IllegalStateException("해당하는 센터가 존재하지 않습니다."));
+
+        if(!center.getDirector().getId().equals(userId)) {
+            throw new IllegalStateException("해당 센터의 소유자가 아닙니다.");
+        }
 
         Class newClass = Class.createClass(classRequest.getName(), center);
         classRepository.save(newClass);
@@ -33,7 +37,7 @@ public class ClassServiceImpl implements ClassService {
 
     @Transactional
     @Override
-    public Class updateClass(Long classId, ClassRequest classRequest) {
+    public Class updateClass(String userId, Long classId, ClassRequest classRequest) {
         Class newClass = classRepository.findById(classId)
                 .orElseThrow(() -> new IllegalStateException("해당 반이 존재하지 않습니다."));
 
@@ -42,26 +46,32 @@ public class ClassServiceImpl implements ClassService {
     }
 
     @Override
-    public List<Class> getClassListByCenter(Long centerId) {
+    public List<Class> getClassListByCenter(String userId, Long centerId) {
         Center center = centerRepository.findById(centerId)
                 .orElseThrow(() -> new IllegalStateException("해당하는 센터가 존재하지 않습니다."));
+
+        if(!center.getDirector().getId().equals(userId)) {
+            throw new IllegalStateException("해당 센터의 소유자가 아닙니다.");
+        }
 
         List<Class> classList = classRepository.findAllByCenterId(center.getId());
         return classList;
     }
 
-    @Override
-    public List<Class> getClassList() {
-        List<Class> classList = classRepository.findAll();
-        return classList;
-    }
+//    @Override
+//    public List<Class> getClassList(String userId) {
+//        List<Class> classList = classRepository.findAll();
+//        return classList;
+//    }
 
     @Transactional
     @Override
-    public void deleteClass(Long classId) {
-        classRepository.findById(classId)
-                        .orElseThrow(() -> new IllegalStateException("해당 반이 존재하지 않습니다."));
+    public boolean deleteClass(String userId, Long classId) {
 
-        classRepository.deleteById(classId);
+        if(classRepository.findById(classId).isPresent()) {
+            classRepository.deleteById(classId);
+            return true;
+        }
+        throw new IllegalStateException("해당하는 반이 존재하지 않습니다.");
     }
 }
